[NEW] Content-Security-Policy for inline scripts (#20724)
