+++ .meteor/packages
@@ -89,3 +89,4 @@ rocketchat:i18n
 rocketchat:postcss
 dandv:caret-position
 facts-base@1.0.1
+browser-policy
