+++ package.json
@@ -65,20 +65,34 @@
 		"@types/chai": "^4.2.12",
 		"@types/chai-spies": "^1.0.1",
 		"@types/clipboard": "^2.0.1",
+		"@types/dompurify": "^2.2.1",
 		"@types/ejson": "^2.1.2",
+		"@types/fibers": "^3.1.0",
+		"@types/imap": "^0.8.33",
 		"@types/jsdom": "^16.2.6",
 		"@types/less": "^3.0.2",
+		"@types/mailparser": "^3.0.0",
+		"@types/marked": "^1.2.2",
 		"@types/meteor": "^1.4.49",
+		"@types/mkdirp": "^1.0.1",
 		"@types/mocha": "^8.0.3",
 		"@types/mock-require": "^2.0.0",
 		"@types/moment-timezone": "^0.5.30",
 		"@types/mongodb": "^3.5.26",
 		"@types/node": "^12.20.10",
+		"@types/nodemailer": "^6.4.0",
+		"@types/parseurl": "^1.3.1",
 		"@types/react": "^17.0.11",
 		"@types/react-dom": "^16.9.13",
 		"@types/rewire": "^2.5.28",
 		"@types/semver": "^7.3.3",
+		"@types/sharp": "^0.28.0",
+		"@types/string-strip-html": "^5.0.0",
 		"@types/toastr": "^2.1.38",
+		"@types/underscore.string": "0.0.38",
+		"@types/use-subscription": "^1.0.0",
+		"@types/xml-crypto": "^1.4.1",
+		"@types/xmldom": "^0.1.30",
 		"@typescript-eslint/eslint-plugin": "^2.34.0",
 		"@typescript-eslint/parser": "^2.34.0",
 		"acorn": "^6.4.1",
@@ -152,19 +166,6 @@
 		"@rocket.chat/string-helpers": "^0.26.0",
 		"@rocket.chat/ui-kit": "^0.26.0",
 		"@slack/client": "^4.12.0",
-		"@types/dompurify": "^2.2.1",
-		"@types/fibers": "^3.1.0",
-		"@types/imap": "^0.8.33",
-		"@types/mailparser": "^3.0.0",
-		"@types/marked": "^1.2.2",
-		"@types/mkdirp": "^1.0.1",
-		"@types/nodemailer": "^6.4.0",
-		"@types/sharp": "^0.28.0",
-		"@types/string-strip-html": "^5.0.0",
-		"@types/underscore.string": "0.0.38",
-		"@types/use-subscription": "^1.0.0",
-		"@types/xml-crypto": "^1.4.1",
-		"@types/xmldom": "^0.1.30",
 		"adm-zip": "RocketChat/adm-zip",
 		"agenda": "^3.1.0",
 		"apn": "2.2.0",
