+++ typings.d.ts
@@ -19,3 +19,9 @@ declare module 'meteor/konecty:user-presence' {
 declare const Package: {
 	'disable-oplog': object;
 };
+
+declare module 'meteor/meteorhacks:inject-initial' {
+	namespace Inject {
+		function rawBody(key: string, value: string): void;
+	}
+}
